requirements.txt
Orange3
beautifulsoup4
bs4
cosine-similarity
fastapi
gradio
httpx
matplotlib
numpy
oaipmh
opencv-python
openai
pandas
pdf2image
pyfuseki
pytesseract
scikit-learn
sparqlwrapper
